pyproject.toml
[project]
name = "menulens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.12.1",
    "anthropic==0.25.1",
    "bcrypt==4.1.1",
    "black==23.11.0",
    "fastapi==0.104.1",
    "flake8==6.1.0",
    "httpx==0.25.1",
    "mypy==1.7.1",
    "openai==1.3.5",
    "passlib[bcrypt]==1.7.4",
    "pillow==10.1.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "sqlalchemy==2.0.23",
    "streamlit==1.29.0",
    "tavily-python==0.3.0",
    "uvicorn[standard]==0.24.0",
]
